move models and repository interfaces in domain layer
