--- api/repositories/repository.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-class Neighbour(BaseModel):
-    repo: str
-    stargazers: list[str]
