+++ api/routers/repos.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends, HTTPException
-from api.repositories.repository import Neighbour
 
+from api.domain.models import Neighbour
 from api.repositories.user import get_current_active_user
 from api.usecases.get_starneighbours import get_starneighbours as get_starneighbours_usecase
