+++ api/usecases/get_starneighbours.py
@@ -1,6 +1,7 @@
-from api.repositories.repository import Neighbour
 import pandas as pd
 import requests
+
+from api.domain.models import Neighbour
 
 # TODO: handle business domain errors, create a dedicated injected github service
 # Core business logic of building starneighbour of a given repository
